backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain>=0.3.0
langchain-groq>=0.2.0
langgraph>=0.2.4
langchain-core>=0.3.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic>=2.7.1
httpx>=0.27.0
